complete auth

+++ app/routes/tasks.py
@@ -1,10 +1,17 @@
 from flask import jsonify, request, Blueprint
 from app.models.task_model import Task
 from app import db
+from flask_login import login_required, current_user
+
 
 tasks_bp = Blueprint("tasks",__name__)
 
+@tasks_bp.route("/")
+def first():
+    return "Hello"
+
 @tasks_bp.route("/api/tasks",methods=["GET"])
+@login_required
 def all_tasks():
     tasks = Task.query.all()
     all_tasks = []
@@ -14,6 +21,7 @@ def all_tasks():
     return jsonify(all_tasks)
 
 @tasks_bp.route("/api/tasks/<int:task_id>",methods=["GET"])
+@login_required
 def get_task(task_id):
     task = Task.query.get(task_id)
     if not task:
@@ -21,6 +29,7 @@ def get_task(task_id):
     return jsonify(task.to_dict())
 
 @tasks_bp.route("/api/tasks",methods=["POST"])
+@login_required
 def create_task():
     data = request.get_json()
     if not data:
@@ -45,6 +54,7 @@ def create_task():
     return jsonify({"success":"task added ", "new task":new_task.to_dict()}),201
 
 @tasks_bp.route("/api/tasks/<int:task_id>",methods=["PUT"])
+@login_required
 def update_task(task_id):
     task = Task.query.get(task_id)
     if not task:
@@ -77,10 +87,13 @@ def update_task(task_id):
 
 
 @tasks_bp.route("/api/tasks/<int:task_id>",methods=["DELETE"])
+@login_required
 def delete_task(task_id):
     task = Task.query.get(task_id)
     if not task:
         return jsonify({"error":"task not found"}),404
+    if task.author != current_user:
+        return jsonify({"error": "Unauthorized"}), 403
     db.session.delete(task)
     db.session.commit()
     return jsonify({"success":"task deleted","task":task.to_dict()}),201
